move Layout into hoc for improved folder structure
